+++ themes/twintack2025/woocommerce/emails/admin-new-order.php
@@ -29,10 +29,10 @@ do_action( 'woocommerce_email_header', $email_heading, $email ); ?>
 <?php
 echo $email_improvements_enabled ? '<div class="email-introduction">' : '';
 /* translators: %s: Customer billing full name */
-$text = __( 'You\'ve received the following order from %s:', 'woocommerce' );
+$text = __( 'You have received the following order from %s:', 'woocommerce' );
 if ( $email_improvements_enabled ) {
 	/* translators: %s: Customer billing full name */
-	$text = __( 'You\'ve received a new order from %s:', 'woocommerce' );
+	$text = __( 'You have received a new order from %s:', 'woocommerce' );
 }
 ?>
 <p><?php printf( esc_html( $text ), esc_html( $order->get_formatted_billing_full_name() ) ); ?></p>